+++ app/(root)/(tabs)/games/create-with-group.tsx
@@ -1,0 +1,11 @@
+import { Text, View } from "react-native";
+
+const CreateWithGroup = () => {
+  return (
+    <View>
+      <Text>Create Game With Group</Text>
+    </View>
+  );
+};
+
+export default CreateWithGroup;
