+++ app/(root)/(tabs)/games/create-without-group.tsx
@@ -1,0 +1,11 @@
+import { Text, View } from "react-native";
+
+const CreateWithoutGroup = () => {
+  return (
+    <View>
+      <Text>Create Game Without Group</Text>
+    </View>
+  );
+};
+
+export default CreateWithoutGroup;
