+++ supabase/config.toml
@@ -317,3 +317,14 @@ entrypoint = "./functions/start-game/index.ts"
 # Specifies static files to be bundled with the function. Supports glob patterns.
 # For example, if you want to serve static HTML pages in your function:
 # static_files = [ "./functions/start-game/*.html" ]
+
+[functions.get-next-restaurant-for-game]
+enabled = true
+verify_jwt = true
+import_map = "./functions/get-next-restaurant-for-game/deno.json"
+# Uncomment to specify a custom file path to the entrypoint.
+# Supported file extensions are: .ts, .js, .mjs, .jsx, .tsx
+entrypoint = "./functions/get-next-restaurant-for-game/index.ts"
+# Specifies static files to be bundled with the function. Supports glob patterns.
+# For example, if you want to serve static HTML pages in your function:
+# static_files = [ "./functions/get-next-restaurant-for-game/*.html" ]
